+++ Makefile
@@ -13,6 +13,7 @@ $(NAME): developing
 	src/lexical_analyzer/*.c \
 	src/syntax_analyzer/*.c \
 	src/environment_variables/*.c \
+	src/character_list/*.c \
 	-Ilib/libft/inc/ -Llib/libft/lib -lft -lreadline -o minishell -g -Wall -Wextra
 
 developing:
