+++ src/character_list/char_list_add_char.c
@@ -1,65 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   char_list.c                                        :+:      :+:    :+:   */
+/*   char_list_add_char.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/07 20:34:06 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/08 17:02:45 by Philip           ###   ########.fr       */
+/*   Created: 2024/04/09 10:10:57 by Philip            #+#    #+#             */
+/*   Updated: 2024/04/09 10:15:08 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "char_list.h"
 #include "libft.h"
-#include <stddef.h>
-#include <stdlib.h>
 
 static t_char_list	*char_list_last_node(t_char_list *list);
-
-void	char_list_free_and_null(t_char_list **list)
-{
-	t_char_list	*node;
-	t_char_list	*next;
-	
-	if (!list || !(*list))
-		return ;
-	node = *list;
-	while (node)
-	{
-		next = node->next;
-		free(node);
-		node = next;
-	}
-	*list = NULL;
-}
-
-char	*char_list_to_str(t_char_list *list)
-{
-	size_t	len;
-	size_t	i;
-	t_char_list	*node;
-	char	*str;
-
-	if (!list)
-		return (NULL);
-	node = list;
-	len = 0;
-	while (node)
-	{
-		node = node->next;
-		len++;
-	}
-	str = (char *)ft_calloc(len + 1, sizeof(char));
-	node = list;
-	i = 0;
-	while (node)
-	{
-		str[i++] = node->c;
-		node = node->next;
-	}
-	return (str);
-}
 
 void	char_list_add_char(t_char_list **list, char c)
 {
@@ -71,15 +25,6 @@ void	char_list_add_char(t_char_list **list, char c)
 		*list = new_node;
 	else
 		char_list_last_node(*list)->next = new_node;
-}
-
-void	char_list_add_str(t_char_list **list, char *str)
-{
-	while (*str)
-	{
-		char_list_add_char(list, *str);
-		str++;
-	}
 }
 
 static t_char_list	*char_list_last_node(t_char_list *list)
