+++ src/expand_arguments.c
@@ -6,13 +6,13 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 17:07:24 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/09 02:47:13 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/09 10:17:45 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cmd_list.h"
 #include "environment_variables/env.h"
-#include "char_list.h"
+#include "character_list/char_list.h"
 #include "free_and_null.h"
 #include "libft.h"
 #include <stddef.h>
